+++ ios-plugin/src/ZendeskSDK.framework/Headers/ZDKCommentInputViewController.h
@@ -22,9 +22,9 @@
 
 @interface ZDKCommentInputViewController : UIViewController
 
-@property (nonatomic) ZDKCommentInputView *commentInputView;
+@property (nonatomic, strong) ZDKCommentInputView *commentInputView;
 
-- (instancetype) initWithRequest:(ZDKRequest*)request;
+- (instancetype)initWithRequest:(ZDKRequest *)request isAttachmentEnabled:(BOOL)isAttachmentEnabled;
 
 - (CGFloat) preferredHeight;