+++ ios-plugin/src/ZendeskSDK/ZDKImageViewerViewController.h
@@ -15,10 +15,12 @@
  */
 
 #import <UIKit/UIKit.h>
-#import <ZendeskSDK/ZendeskSDK.h>
-
+#import "ZDKUIViewController.h"
+#import "ZDKRotationForwarding.h"
 
 @interface ZDKImageViewerViewController : ZDKUIViewController<UIScrollViewDelegate>
+
+@property (nonatomic, assign) id<ZDKRotationForwarding> rotationEventDelegate;
 
 - (instancetype) initWithImage:(UIImage*)image;