+++ sample/ZendeskExample.cs
@@ -4,24 +4,35 @@ using System.Collections;
 using ZendeskSDK;
 
 // <summary>
-// A simple example demonstrating Zendek configuration and displaying the
+// A simple example demonstrating Zendesk configuration and displaying the
 // Help Center, Requests and Rate My App views.
 // </summary>
 public class ZendeskExample: MonoBehaviour
 {
 
-	/** initialize zendesk in the Awake() method of the GameObject a script of yours is attached to */
+	// <summary>
+    // This shows you how to initialize the SDK and set an identity. These actions are required
+    // before you can interact with the SDK.
+    // </summary>
 	void Awake() {
-		ZendeskSDK.ZDKConfig.Initialize (gameObject); // DontDestroyOnLoad automatically called on your supplied gameObject
-		ZendeskSDK.ZDKConfig.AuthenticateJwtUserIdentity ("<UserId>");
+
+        // Initializes the SDK using your subdomain, application ID, and oauth client ID. Related info:
+        // Android: https://developer.zendesk.com/embeddables/docs/android/initialize_sdk
+        // iOS: https://developer.zendesk.com/embeddables/docs/ios/initialize_sdk
+		ZendeskSDK.ZDKConfig.Initialize (gameObject, "https://{subdomain}.zendesk.com", "{applicationId}", "{oauthClientId}");
+
+		// Set an anonymous identity. Releated info:
+		// Android: https://developer.zendesk.com/embeddables/docs/android/set_an_identity
+		// iOS: https://developer.zendesk.com/embeddables/docs/ios/set_an_identity
+		ZendeskSDK.ZDKConfig.AuthenticateAnonymousIdentity();
+
+		// If you use JWT identities, then comment out the AuthenticateAnonymousIdentity line above, and uncomment this line.
+		//ZendeskSDK.ZDKConfig.AuthenticateJwtUserIdentity ("MyTestID");
 	}
-	
-	/** must include this method for android to behave properly */
-	void OnApplicationPause(bool pauseStatus) {
-		ZendeskSDK.ZDKConfig.OnApplicationPause (pauseStatus);
-	}
-	
-	/** must include this method for any zendesk callbacks to work */
+
+	// <summary>
+    // must include this method for any zendesk callbacks to work
+    // </summary>
 	void OnZendeskCallback(string results) {
 		ZDKConfig.CallbackResponse (results);
 	}
@@ -34,6 +45,10 @@ public class ZendeskExample: MonoBehaviour
 
 	}
 
+    // <summary>
+    // This shows you how to add a few buttons which can be used to launch different parts of the
+    // SDK. You must have previously initialized the SDK and set an identity as shown in Awake()
+    // </summary>
 	void OnGUI() {
 		GUI.matrix = Matrix4x4.Scale (new Vector3 (5, 5, 5));
 
@@ -43,10 +58,6 @@ public class ZendeskExample: MonoBehaviour
 
 		if (GUILayout.Button ("Request Creation")) {
 			ZendeskSDK.ZDKRequests.ShowRequestCreation ();
-		}
-
-		if (GUILayout.Button ("Requests List")) {
-			ZendeskSDK.ZDKRequests.ShowRequestList ();
 		}
 
 		if (GUILayout.Button ("Rate My App")) {