+++ CHANGELOG.md
@@ -1,10 +1,34 @@
-# v.1.4.1.1
+# v1.7.0.1
+
+## Support SDK versions
+
+* Updates native iOS Support SDK to [1.7.5.1](https://developer.zendesk.com/embeddables/docs/ios/version_information).
+* Updates native Android Support SDK to [1.7.2.1](https://developer.zendesk.com/embeddables/docs/android/version_information)
+
+[Learn more](https://support.zendesk.com/hc/en-us/articles/224252968-Introducing-Support-SDK-v1-7-A-new-Help-Center-experience-and-options) about the new Help Center experience.
+
+## Push notifications
+
+Version 1.4.1.1 of the Unity plugin contained end-to-end support for push notifications, and we bundled UrbanAirship's SDK to accomplish this.
+
+In version 1.7.0.1 we removed the client-side push libraries to be more in line with how push is handled in the native Support SDKs. If you wish to use push notifications you will need to determine the client id or channel id of your users' device, and then pass the identifier to `ZendeskSDK.ZDKPush.EnableWithIdentifier`. More details of push notifications are available below:
+
+* [Android web-hook application integration](https://developer.zendesk.com/embeddables/docs/android/handle_push_notifications_wh#application-integration)
+* [Android UrbanAirship application integration](https://developer.zendesk.com/embeddables/docs/android/handle_push_notifications_ua#application-integration)
+* [iOS web-hook application integration](https://developer.zendesk.com/embeddables/docs/ios/handle_push_notifications_wh#application-integration)
+* [iOS UrbanAirship application integration](https://developer.zendesk.com/embeddables/docs/ios/handle_push_notifications_ua#application-integration)
+
+## Entry points
+
+Version 1.7.0.1 supports Help Center, or Create Request as the main entry points. Version 1.4.1.1 allowed the request list as an entry point.
+
+# v1.4.1.1
 * Fix crashes for iOS7 and below when initializing the SDK.
 * Fix issue of incorrectly mapping ticket email to ticket Subject, ticket Subject to ticket description, and ticket description to ticket email.
 
-# v.1.4.0.1
+# v1.4.0.1
 
-* **Native SDKs updated: plugin now uses Zendesk Android SDK v1.4.2.2 and iOS SDK v.1.4.2.1. Please see the changelogs for these products to read about any under-the-hood improvements.**
+* **Native SDKs updated: plugin now uses Zendesk Android SDK v1.4.2.2 and iOS SDK v1.4.2.1. Please see the changelogs for these products to read about any under-the-hood improvements.**
     * https://developer.zendesk.com/embeddables/docs/android/release_notes
     * https://developer.zendesk.com/embeddables/docs/ios/release_notes
 * ZDKConfig changes:
@@ -14,7 +38,7 @@
     * Method removed: `ShowRequestCreation(ZDKRequestCreationConfig config)` - Use `ZDKConfig.SetContactConfiguration` instead.
 * ZDKHelpCenter changes:
     * Method removed: `SetRequestSubject(String reqSub)` - Use `ZDKConfig.SetContactConfiguration` instead.
-*  Synchronise default request subject across both Android and iOS.
+*  Synchronize default request subject across both Android and iOS.
 *  Fixes to push notifications on Android via UrbanAirship.
 *  Android permissions added, for adding and viewing attachments in tickets:
     * READ_EXTERNAL_STORAGE
@@ -23,7 +47,7 @@
 *  Fixes to push notifications on iOS.
 *  `ZendeskTester` updated with more examples.
 
-# v.1.3
+# v1.3
 
 * **Native SDKs updated: plugin now uses Zendesk Android SDK v1.4.1.1 and iOS SDK v1.4.1.4. Please see the changelogs for these products to read about any under-the-hood improvements.**
     * https://developer.zendesk.com/embeddables/docs/android/release_notes
@@ -45,17 +69,17 @@
     * ZDKConfig.SetDebugLoggingiOS removed, replaced with ZDKLogger.Enable and ZDKLogger.SetLogLevelIOS
     * new methods added: SetCoppaEnabled, SetUserLocale
 
-# v.1.2
+# v1.2
 
 * Added ViewArticle method to the Help Center Provider.
 * Added Alpha color and background color for Rate My App view.
 * Update README to state that Unity 5.0 is required.
 
-# v.1.1
+# v1.1
 
 * Android - Fix upload provider to work with image files
 * Fix conflicts with Chartboost Unity Plugin
 
-# v.1.0
+# v1.0
 
 * Initial release